some fixes

+++ services/university/models/grade_statistic_response.py
@@ -5,4 +5,4 @@ class GradeStatisticResponse(BaseModel):
     count: int = Field(ge=0)
     min: int | None
     max: int | None
-    avg: int | None
+    avg: float | None
